(refactor) Refining the Claims File Structure and responsiveness of the  claims page using the Openmrs guidelines
