+++ packages/esm-billing-app/src/claims/claims-wrap/claims-main.scss
@@ -10,7 +10,7 @@
   flex-direction: column;
 }
 
-@media (min-width: 768px) {
+:global(.omrs-breakpoint-lt-desktop){
   .mainContainer {
     flex-direction: column;
   }
