+++ packages/esm-billing-app/src/claims/dashboard/breakdown/claims-breakdown.scss
