+++ packages/esm-billing-app/src/claims/dashboard/header/claims-header.scss
