+++ packages/esm-billing-app/src/claims/dashboard/table/claims-table.scss
@@ -40,7 +40,7 @@
   justify-content: flex-start;
 }
 
-@media (min-width: 768px) {
+:global(.omrs-breakpoint-lt-desktop){
   .claimContainer {
     width: 50%;
   }
